+++ lib/custom_accordion.dart
@@ -7,8 +7,10 @@ class CustomAccordion extends StatefulWidget {
   CustomAccordion({
     super.key,
     required this.title,
+    this.subTitle,
     this.headerBackgroundColor,
     this.titleStyle,
+    this.subTitleStyle,
     this.toggleIconOpen,
     this.toggleIconClose,
     this.headerIconColor,
@@ -18,11 +20,14 @@ class CustomAccordion extends StatefulWidget {
     this.widgetItemsPadding,
     // this.actionButton,
     this.showContent = false,
+    this.disableToggle = false,
   });
 
   final String title;
+  final String? subTitle;
   final Color? headerBackgroundColor;
   final TextStyle? titleStyle;
+  final TextStyle? subTitleStyle;
   final IconData? toggleIconOpen;
   final IconData? toggleIconClose;
   final Color? headerIconColor;
@@ -32,6 +37,7 @@ class CustomAccordion extends StatefulWidget {
   final EdgeInsetsGeometry? widgetItemsPadding;
   // final Widget? actionButton;
   bool showContent;
+  bool disableToggle;
 
   @override
   State<CustomAccordion> createState() => _CustomAccordionState();
@@ -47,35 +53,34 @@ class _CustomAccordionState extends State<CustomAccordion> {
       color: widget.backgroundColor ?? Theme.of(context).cardColor,
       child: Column(children: [
         ListTile(
-          onTap: () {
+          onTap: widget.disableToggle ? null : () {
             setState(() {
               widget.showContent = !(widget.showContent);
             });
           },
           tileColor:
               widget.headerBackgroundColor ?? Theme.of(context).splashColor,
-          title: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Expanded(
-                child: Text(
-                  widget.title.toUpperCase(),
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                  style: widget.titleStyle ??
-                      Theme.of(context).textTheme.titleLarge,
-                ),
-              ),
-              // widget.actionButton ?? const SizedBox()
-            ],
+          title: Text(
+            widget.title.toUpperCase(),
+            maxLines: 1,
+            overflow: TextOverflow.ellipsis,
+            style: widget.titleStyle ??
+                Theme.of(context).textTheme.titleLarge,
           ),
-          trailing: IconButton(
+          subtitle: widget.subTitle != null ? Text(
+            widget.subTitle ?? '',
+            maxLines: 1,
+            overflow: TextOverflow.ellipsis,
+            style: widget.subTitleStyle ??
+                Theme.of(context).textTheme.titleSmall,
+          ) : null,
+          trailing: widget.disableToggle ? null : IconButton(
             icon: Icon(
               widget.showContent
                   ? (widget.toggleIconClose ?? Icons.arrow_drop_up)
                   : (widget.toggleIconOpen ?? Icons.arrow_drop_down),
               color:
-                  widget.headerIconColor ?? Theme.of(context).iconTheme.color,
+              widget.headerIconColor ?? Theme.of(context).iconTheme.color,
             ),
             tooltip: 'Toggle Accordion',
             splashRadius: 20,
@@ -86,7 +91,17 @@ class _CustomAccordionState extends State<CustomAccordion> {
             },
           ),
         ),
-        (widget.showContent)
+        widget.disableToggle
+          ? Container(
+          padding: widget.widgetItemsPadding ?? const EdgeInsets.all(10),
+          child: widget.widgetItems ??
+              Center(
+                  child: Text(
+                    'No Widget available',
+                    style: Theme.of(context).textTheme.bodyMedium,
+                  )),
+        )
+          : (widget.showContent)
             ? Container(
                 padding: widget.widgetItemsPadding ?? const EdgeInsets.all(10),
                 child: widget.widgetItems ??
@@ -96,7 +111,7 @@ class _CustomAccordionState extends State<CustomAccordion> {
                       style: Theme.of(context).textTheme.bodyMedium,
                     )),
               )
-            : Container()
+            : const SizedBox(),
       ]),
     );
   }