+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Added sub title under accordion heading
+* handled toggle functionality as disableToggle
+
 ## 0.0.2
 
 * Demo preview added in documentation.